+++ lib/frise.rb
@@ -1,5 +1,7 @@
 # frozen_string_literal: true
 
+require 'liquid/liquid'
+
 require 'frise/defaults_loader'
 require 'frise/loader/lazy'
 require 'frise/loader'