+++ lib/liquid/liquid.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+require 'liquid'
+
+require_relative 'json_filter'
+
+Liquid::Template.register_filter(JsonFilter)