+++ CHANGELOG.md
@@ -4,6 +4,9 @@
   - Liquid parsing is now strict ([#29](https://github.com/velocidi/frise/pull/29)).
   - Update liquid gem to `5.3.0` ([#29](https://github.com/velocidi/frise/pull/29)).
 
+- New features
+  - Add `json` liquid filter ([#28](https://github.com/velocidi/frise/pull/28)).
+
 ### 0.5.1 (March 28, 2022)
 
 - Breaking changes